+++ src/Programs/QHE/QHEFermionsCoulomb.cc
@@ -98,7 +98,7 @@ int main(int argc, char** argv)
   int NbrEigenvalue = NbrEigenvaluesOption.GetInteger();
   int NbrFermions = NbrFermionOption.GetInteger();
   int LzMax = LzMaxOption.GetInteger();
-  int Memory = MemoryOption.GetInteger() << 20;
+  long Memory = MemoryOption.GetInteger() << 20;
   bool DeltaFlag = DeltaOption.GetBoolean();
   double CoulombRatio = CoulombRatioOption.GetDouble();
   int InitialLz = InitialLzOption.GetInteger();
